% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_factory.R
\name{make_server}
\alias{make_server}
\title{create a shiny server to be wired up to sliders}
\usage{
make_server(mod, shiny_param_list)
}
\arguments{
\item{mod}{model parameter}

\item{shiny_param_list}{list of param and initial values}
}
\description{
create a shiny server to be wired up to sliders
}
\details{
shiny_param_list comes from as.list(param(mod)) and any
prior filtering. Any passed in will expect to be modifying
the values from the slider inputs
}

