% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_factory.R
\name{make_ui}
\alias{make_ui}
\title{ui factory function}
\usage{
make_ui(param_list, params_per_tab = 5)
}
\arguments{
\item{param_list}{list of params to turn into sliders}

\item{params_per_tab}{number of parameters per tab}
}
\description{
ui factory function
}

