% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgexplore.R
\name{mrgexplore}
\alias{mrgexplore}
\title{launch shiny app to explore influence of parameter values}
\usage{
mrgexplore(mod)
}
\arguments{
\item{mod}{mrgsolve model}
}
\description{
launch shiny app to explore influence of parameter values
}

